refactor(core): extract lifecycle and extraction helpers
