+++ packages/remnic-core/src/types.ts
@@ -5,6 +5,7 @@ import type { WearablesConfig } from "./wearables/types.js";
 import type { ExtractionLivenessConfig } from "./extraction-liveness.js";
 import type { ReplicaPeersConfig } from "./replica-peers-config.js";
 
+import type { BufferTurnOwner } from "./buffer-turn-helpers.js";
 export type ReasoningEffort = "none" | "low" | "medium" | "high";
 export type TriggerMode = "smart" | "every_n" | "time_based";
 export type SignalLevel = "none" | "low" | "medium" | "high";
@@ -2646,19 +2647,12 @@ export interface RelevanceFeedback {
 export interface SourceConnectorProvenance {
   sourceConnector?: string;
 }
-
-export interface BufferTurn extends SourceConnectorProvenance {
+export interface BufferTurn extends SourceConnectorProvenance, BufferTurnOwner {
   role: "user" | "assistant";
   content: string;
   timestamp: string;
   sourceValidAt?: string;
   sessionKey?: string;
-  /**
-   * Server-resolved principal that owns this session at the authenticated
-   * observe/write boundary. Persisted with the buffered turn so lifecycle
-   * flushes do not have to infer ownership from a client-chosen session key.
-   */
-  sessionOwnerPrincipal?: string;
   logicalSessionKey?: string;
   providerThreadId?: string | null;
   turnFingerprint?: string;
